+++ test/e4-newMesh.js
@@ -55,6 +55,43 @@ describe('e4-newMesh', function () {
 
   });
 
+  it('starts a mesh with deferListen true and no loader', function (done) {
+
+    var configDeferredListen = {
+      name: "startupProxiedDifferentPort",
+      port: 55001,
+      deferListen: true
+    };
+
+    Mesh
+      .create(configDeferredListen, function (e, created) {
+
+        doRequest('/ping', function (data) {
+
+          expect(data).to.be('pong');
+          created.stop(done);
+        },55001);
+      });
+  });
+
   //require('benchmarket').stop();
 
 });
+
+function doRequest(path, callback, port) {
+
+  var request = require('request');
+
+  if (!port) port = 55000;
+
+  if (path[0] != '/')
+    path = '/' + path;
+
+  var options = {
+    url: 'http://127.0.0.1:' + port.toString() + path,
+  };
+
+  request(options, function (error, response, body) {
+    callback(body);
+  });
+}